Moved skeleton building into dedicated class.
