+++ lionelmssq/prediction.py
@@ -1,21 +1,24 @@
 from dataclasses import dataclass
-from itertools import chain, groupby
+# from itertools import chain, groupby
 from pathlib import Path
-from typing import List, Optional, Self, Set, Tuple
+from typing import List, Self
 from lionelmssq.common import Side
 from lionelmssq.linear_program import LinearProgramInstance
 from lionelmssq.mass_table import DynamicProgrammingTable
 from lionelmssq.masses import UNIQUE_MASSES, EXPLANATION_MASSES, MATCHING_THRESHOLD
-from lionelmssq.mass_explanation import explain_mass
+# from lionelmssq.mass_explanation import explain_mass
 import polars as pl
 from loguru import logger
 
+from lionelmssq.skeleton_building import SkeletonBuilder, \
+    calculate_diff_dp, calculate_diff_errors
 
-@dataclass
-class TerminalFragment:
-    index: int  # fragment index
-    min_end: int  # minimum length of fragment (negative for end fragments)
-    max_end: int  # maximum length of fragment (negative for end fragments)
+
+# @dataclass
+# class TerminalFragment:
+#     index: int  # fragment index
+#     min_end: int  # minimum length of fragment (negative for end fragments)
+#     max_end: int  # maximum length of fragment (negative for end fragments)
 
 
 @dataclass
@@ -88,37 +91,37 @@ class Predictor:
         self.fragment_masses = dict()
         self.dp_table = dp_table
 
-    def build_skeleton(
-        self, modification_rate
-    ) -> Tuple[
-        List[Set[str]],
-        List[TerminalFragment],
-        List[TerminalFragment],
-        List[int],
-        List[int],
-    ]:
-        skeleton_seq_start, start_fragments, invalid_start_fragments = (
-            self._predict_skeleton(Side.START, modification_rate)
-        )
-
-        print("Skeleton sequence start = ", skeleton_seq_start)
-
-        skeleton_seq_end, end_fragments, invalid_end_fragments = self._predict_skeleton(
-            Side.END, modification_rate
-        )
-        print("Skeleton sequence end = ", skeleton_seq_end)
-
-        skeleton_seq = self._align_skeletons(skeleton_seq_start, skeleton_seq_end)
-
-        print("Skeleton sequence = ", skeleton_seq)
-
-        return (
-            skeleton_seq,
-            start_fragments,
-            end_fragments,
-            invalid_start_fragments,
-            invalid_end_fragments,
-        )
+    # def build_skeleton(
+    #     self, modification_rate
+    # ) -> Tuple[
+    #     List[Set[str]],
+    #     List[TerminalFragment],
+    #     List[TerminalFragment],
+    #     List[int],
+    #     List[int],
+    # ]:
+    #     skeleton_seq_start, start_fragments, invalid_start_fragments = (
+    #         self._predict_skeleton(Side.START, modification_rate)
+    #     )
+    #
+    #     print("Skeleton sequence start = ", skeleton_seq_start)
+    #
+    #     skeleton_seq_end, end_fragments, invalid_end_fragments = self._predict_skeleton(
+    #         Side.END, modification_rate
+    #     )
+    #     print("Skeleton sequence end = ", skeleton_seq_end)
+    #
+    #     skeleton_seq = self._align_skeletons(skeleton_seq_start, skeleton_seq_end)
+    #
+    #     print("Skeleton sequence = ", skeleton_seq)
+    #
+    #     return (
+    #         skeleton_seq,
+    #         start_fragments,
+    #         end_fragments,
+    #         invalid_start_fragments,
+    #         invalid_end_fragments,
+    #     )
 
     def predict(self, modification_rate: float = 0.5) -> Prediction:
         # Adapt individual modification rates to universal one
@@ -161,6 +164,17 @@ class Predictor:
             self._reduce_alphabet()
         )
 
+        skeleton_builder = SkeletonBuilder(
+            fragment_masses=self.fragment_masses,
+            fragments_side=self.fragments_side,
+            mass_diffs=self.mass_diffs,
+            explanations=self.explanations,
+            seq_len=self.seq_len,
+            matching_threshold=self.matching_threshold,
+            mass_tags=self.mass_tags,
+            dp_table=self.dp_table,
+        )
+
         # Now we build the skeleton sequence from both sides and align them to get the final skeleton sequence!
         (
             skeleton_seq,
@@ -168,7 +182,7 @@ class Predictor:
             end_fragments,
             invalid_start_fragments,
             invalid_end_fragments,
-        ) = self.build_skeleton(modification_rate)
+        ) = skeleton_builder.build_skeleton(modification_rate)
 
         # TODO: If the tags are considered in the LP at the end, then most of the following code will become obsolete!
 
@@ -379,24 +393,24 @@ class Predictor:
 
         self.fragment_masses[side] = side_fragments + start_end_fragments
 
-    def _align_skeletons(self, skeleton_seq_start, skeleton_seq_end) -> List[Set[str]]:
-        # While building the ladder it may happen that things are unambiguous from one side, but not from the other!
-        # In that case, we should consider the unambiguous side as the correct one! If the intersection is empty, then we can consider the union of the two!
-
-        # Align the skeletons of the start and end fragments to get the final skeleton sequence!
-        # Wherever there is no ambiguity, that nucleotide is preferrentially considered!
-
-        skeleton_seq = [set() for _ in range(self.seq_len)]
-        for i in range(self.seq_len):
-            skeleton_seq[i] = skeleton_seq_start[i].intersection(skeleton_seq_end[i])
-            if not skeleton_seq[i]:
-                skeleton_seq[i] = skeleton_seq_start[i].union(skeleton_seq_end[i])
-
-        # TODO: Its more complicated, since if two positions are ambiguous,
-        #  they are not independent. If one nucleotide is selected this way,
-        #  then the same nucleotide cannot be selected in the other position!
-
-        return skeleton_seq
+    # def _align_skeletons(self, skeleton_seq_start, skeleton_seq_end) -> List[Set[str]]:
+    #     # While building the ladder it may happen that things are unambiguous from one side, but not from the other!
+    #     # In that case, we should consider the unambiguous side as the correct one! If the intersection is empty, then we can consider the union of the two!
+    #
+    #     # Align the skeletons of the start and end fragments to get the final skeleton sequence!
+    #     # Wherever there is no ambiguity, that nucleotide is preferrentially considered!
+    #
+    #     skeleton_seq = [set() for _ in range(self.seq_len)]
+    #     for i in range(self.seq_len):
+    #         skeleton_seq[i] = skeleton_seq_start[i].intersection(skeleton_seq_end[i])
+    #         if not skeleton_seq[i]:
+    #             skeleton_seq[i] = skeleton_seq_start[i].union(skeleton_seq_end[i])
+    #
+    #     # TODO: Its more complicated, since if two positions are ambiguous,
+    #     #  they are not independent. If one nucleotide is selected this way,
+    #     #  then the same nucleotide cannot be selected in the other position!
+    #
+    #     return skeleton_seq
 
     def _collect_diffs(self, side: Side) -> None:
         masses = self.fragment_masses[side]
@@ -405,13 +419,13 @@ class Predictor:
             masses[i] - masses[i - 1] for i in range(1, len(masses))
         ]
         self.mass_diffs_errors[side] = [
-            self._calculate_diff_errors(
+            calculate_diff_errors(
                 self.mass_tags[side],
                 masses[0] + self.mass_tags[side],
                 self.matching_threshold,
             )
         ] + [
-            self._calculate_diff_errors(
+            calculate_diff_errors(
                 masses[i] + self.mass_tags[side],
                 masses[i - 1] + self.mass_tags[side],
                 self.matching_threshold,
@@ -419,13 +433,13 @@ class Predictor:
             for i in range(1, len(masses))
         ]
 
-    def _calculate_diff_errors(self, mass1, mass2, threshold) -> float:
-        retval = threshold * ((mass1**2 + mass2**2) ** 0.5) / abs(mass1 - mass2)
-        # Constrain the maximum relative error to 1!
-        # For mass difference very close to zero, the relative error can be very high!
-        if retval > 1:
-            retval = 1.0
-        return retval
+    # def _calculate_diff_errors(self, mass1, mass2, threshold) -> float:
+    #     retval = threshold * ((mass1**2 + mass2**2) ** 0.5) / abs(mass1 - mass2)
+    #     # Constrain the maximum relative error to 1!
+    #     # For mass difference very close to zero, the relative error can be very high!
+    #     if retval > 1:
+    #         retval = 1.0
+    #     return retval
 
     def _collect_singleton_masses(
         self,
@@ -435,28 +449,28 @@ class Predictor:
         )
         self.singleton_masses = set(masses)
 
-    def _calculate_diff_dp(self, diff, threshold, modification_rate):
-        # TODO: Add support for other breakages than 'c/y_c/y'
-        explanation_list = list(
-            [
-                entry
-                for entry in explain_mass(
-                    diff,
-                    dp_table=self.dp_table,
-                    seq_len=self.seq_len,
-                    max_modifications=round(modification_rate * self.seq_len),
-                    threshold=threshold,
-                )
-                if entry.breakage == "c/y_c/y"
-            ][0].explanations
-        )
-        if len(explanation_list) > 0:
-            retval = [
-                Explanation(*explanation_list[i]) for i in range(len(explanation_list))
-            ]
-        else:
-            retval = []
-        return retval
+    # def _calculate_diff_dp(self, diff, threshold, modification_rate):
+    #     # TODO: Add support for other breakages than 'c/y_c/y'
+    #     explanation_list = list(
+    #         [
+    #             entry
+    #             for entry in explain_mass(
+    #                 diff,
+    #                 dp_table=self.dp_table,
+    #                 seq_len=self.seq_len,
+    #                 max_modifications=round(modification_rate * self.seq_len),
+    #                 threshold=threshold,
+    #             )
+    #             if entry.breakage == "c/y_c/y"
+    #         ][0].explanations
+    #     )
+    #     if len(explanation_list) > 0:
+    #         retval = [
+    #             Explanation(*explanation_list[i]) for i in range(len(explanation_list))
+    #         ]
+    #     else:
+    #         retval = []
+    #     return retval
 
     def _collect_diff_explanations(self, modification_rate) -> None:
         diffs = (self.mass_diffs[Side.START]) + (self.mass_diffs[Side.END])
@@ -465,13 +479,20 @@ class Predictor:
             (self.mass_diffs_errors[Side.START]) + (self.mass_diffs_errors[Side.END])
         )
         for diff, diff_error in zip(diffs, diffs_errors):
-            self.explanations[diff] = self._calculate_diff_dp(
-                diff, diff_error, modification_rate
+            # self.explanations[diff] = self._calculate_diff_dp(
+            #     diff, diff_error, modification_rate
+            # )
+            self.explanations[diff] = calculate_diff_dp(
+                diff, diff_error, modification_rate, self.seq_len, self.dp_table
             )
 
         for diff in self.singleton_masses:
-            self.explanations[diff] = self._calculate_diff_dp(
-                diff, self.matching_threshold, modification_rate
+            # self.explanations[diff] = self._calculate_diff_dp(
+            #     diff, self.matching_threshold, modification_rate
+            #
+            self.explanations[diff] = calculate_diff_dp(
+                diff, self.matching_threshold, modification_rate,
+                self.seq_len, self.dp_table
             )
         # TODO: Can make it simpler here by rejecting diff which cannot be explained instead of doing it in the _predict_skeleton function!
 
@@ -493,185 +514,184 @@ class Predictor:
 
         return reduced
 
-    def _predict_skeleton(
-        self,
-        side: Side,
-        modification_rate,
-        skeleton_seq: Optional[List[Set[str]]] = None,
-        fragment_masses=None,
-        candidate_fragments=None,
-    ) -> Tuple[List[Set[str]], List[TerminalFragment], List[int]]:
-        if skeleton_seq is None:
-            skeleton_seq = [set() for _ in range(self.seq_len)]
+    # def _predict_skeleton(
+    #     self,
+    #     side: Side,
+    #     modification_rate,
+    #     skeleton_seq: Optional[List[Set[str]]] = None,
+    #     fragment_masses=None,
+    #     candidate_fragments=None,
+    # ) -> Tuple[List[Set[str]], List[TerminalFragment], List[int]]:
+    #     if skeleton_seq is None:
+    #         skeleton_seq = [set() for _ in range(self.seq_len)]
+    #
+    #     factor = 1 if side == Side.START else -1
+    #
+    #     if not fragment_masses:
+    #         fragment_masses = self.fragment_masses[side]
+    #
+    #     if not candidate_fragments:
+    #         candidate_fragments = (
+    #             self.fragments_side[side].get_column("index").to_list()
+    #         )
+    #
+    #     def get_possible_nucleosides(pos: int, i: int) -> Set[str]:
+    #         return skeleton_seq[pos + factor * i]
+    #
+    #     def is_valid_pos(pos: int, ext: int) -> bool:
+    #         pos = pos + factor * ext
+    #         return (
+    #             # 0 <= pos <= self.seq_len
+    #             0 <= pos < self.seq_len
+    #             if side == Side.START
+    #             # else -(self.seq_len + 1) <= pos < 0
+    #             else -self.seq_len <= pos < 0
+    #         )
+    #
+    #     pos = {0} if side == Side.START else {-1}
+    #
+    #     # We reject some masses/some fragments which are not explained well by mass differences.
+    #     # While iterating through the fragments, the "carry_over_mass" keeps a track of the rejected mass difference.
+    #     # This is added to "diff" to get the next mass_difference.
+    #     carry_over_mass = 0
+    #
+    #     # "last_mass_valid" keeps a track of the last mass which was NOT rejected.
+    #     # This is useful for calculating the difference threshold
+    #     last_mass_valid = 0.0
+    #
+    #     valid_terminal_fragments = []
+    #     invalid_fragments = []
+    #     for fragment_index, (diff, mass) in enumerate(
+    #         zip(self.mass_diffs[side], fragment_masses)
+    #     ):
+    #         diff += carry_over_mass
+    #         assert pos
+    #
+    #         is_valid = True
+    #         if diff in self.explanations:
+    #             explanations = self.explanations.get(diff, [])
+    #         else:
+    #             threshold = self._calculate_diff_errors(
+    #                 last_mass_valid + self.mass_tags[side],
+    #                 last_mass_valid + diff + self.mass_tags[side],
+    #                 self.matching_threshold,
+    #             )
+    #             explanations = self._calculate_diff_dp(
+    #                 diff, threshold, modification_rate
+    #             )
+    #
+    #         # METHOD: if there is only one single nucleoside explanation, we can
+    #         # directly assign the nucleoside if there are tuples, we have to assign a
+    #         # possibility to the current and next position
+    #         # and take care of expressing that both permutations are possible
+    #         # Further, we consider multiple possible start positions, depending on
+    #         # whether the previous explanations had different lengths.
+    #         if explanations:
+    #             next_pos = set()
+    #             if side is Side.START:
+    #                 min_p = min(pos)
+    #                 max_p = max(pos)
+    #             else:
+    #                 min_p = max(pos)
+    #                 max_p = min(pos)
+    #             min_fragment_end = None
+    #             max_fragment_end = None
+    #
+    #             for p in pos:
+    #                 p_specific_explanations = [
+    #                     expl for expl in explanations if is_valid_pos(p, len(expl))
+    #                 ]
+    #                 alphabet_per_expl_len = {
+    #                     expl_len: set(chain(*expls))
+    #                     for expl_len, expls in groupby(p_specific_explanations, len)
+    #                 }
+    #
+    #                 if p_specific_explanations:
+    #                     if p == min_p:
+    #                         min_fragment_end = min_p + factor * min(
+    #                             expl_len for expl_len in alphabet_per_expl_len
+    #                         )
+    #                     elif p == max_p:
+    #                         max_fragment_end = max_p + factor * max(
+    #                             expl_len for expl_len in alphabet_per_expl_len
+    #                         )
+    #
+    #                 # constrain already existing sets in the range of the expl
+    #                 # by the nucs that are given in the explanations
+    #                 for expl_len, alphabet in alphabet_per_expl_len.items():
+    #                     for i in range(expl_len):
+    #                         possible_nucleosides = get_possible_nucleosides(p, i)
+    #
+    #                         if possible_nucleosides.issuperset(alphabet):
+    #                             # If the current explanation sharpens the list of possibilities, clear all
+    #                             # prior possibilities before the new explanation will add the sharpened ones below
+    #                             possible_nucleosides.clear()
+    #
+    #                         for j in alphabet:
+    #                             possible_nucleosides.add(j)
+    #                             # TODO: We need to do this better.
+    #                             # Instead of adding just the letters, we somehow need to keep a track of the possibilities to be able to constrain the LP!
+    #                             # We also then probably need part of the code immediately below!
+    #
+    #                 # add possible follow up positions
+    #                 next_pos.update(
+    #                     p + factor * expl_len for expl_len in alphabet_per_expl_len
+    #                 )
+    #             if max_fragment_end is None:
+    #                 max_fragment_end = min_fragment_end
+    #             if min_fragment_end is None:
+    #                 min_fragment_end = max_fragment_end
+    #             if min_fragment_end is None:
+    #                 # still None => both are None
+    #                 # TODO can we stop early in building the ladder?
+    #                 is_valid = False
+    #         elif (
+    #             abs(diff) <= self.matching_threshold * abs(mass + self.mass_tags[side])
+    #             # Problem! The above approach might blow up if the masses are very close, i.e. diff is very close to zero!
+    #         ):
+    #             if side == Side.START:
+    #                 min_fragment_end = min(pos)
+    #                 max_fragment_end = max(pos)
+    #             else:
+    #                 min_fragment_end = max(pos)
+    #                 max_fragment_end = min(pos)
+    #             next_pos = pos
+    #         else:
+    #             is_valid = False
+    #
+    #         if is_valid:
+    #             valid_terminal_fragments.append(
+    #                 TerminalFragment(
+    #                     index=candidate_fragments[fragment_index],
+    #                     min_end=min_fragment_end,
+    #                     max_end=max_fragment_end,
+    #                 )
+    #             )
+    #             pos = next_pos
+    #             last_mass_valid = mass
+    #             carry_over_mass = 0.0
+    #         else:
+    #             logger.warning(
+    #                 f"Skipping {side} fragment {fragment_index} with observed mass {mass} because no "
+    #                 "explanations are found for the mass difference."
+    #             )
+    #             carry_over_mass = diff
+    #
+    #             # Consider the skipped fragments as internal fragments! Add back the terminal mass to this fragments!
+    #             if fragment_index and candidate_fragments:
+    #                 invalid_fragments.append(candidate_fragments[fragment_index])
+    #
+    #     return skeleton_seq, valid_terminal_fragments, invalid_fragments
 
-        factor = 1 if side == Side.START else -1
-
-        if not fragment_masses:
-            fragment_masses = self.fragment_masses[side]
-
-        if not candidate_fragments:
-            candidate_fragments = (
-                self.fragments_side[side].get_column("index").to_list()
-            )
-
-        def get_possible_nucleosides(pos: int, i: int) -> Set[str]:
-            return skeleton_seq[pos + factor * i]
-
-        def is_valid_pos(pos: int, ext: int) -> bool:
-            pos = pos + factor * ext
-            return (
-                # 0 <= pos <= self.seq_len
-                0 <= pos < self.seq_len
-                if side == Side.START
-                # else -(self.seq_len + 1) <= pos < 0
-                else -self.seq_len <= pos < 0
-            )
-
-        pos = {0} if side == Side.START else {-1}
-
-        # We reject some masses/some fragments which are not explained well by mass differences.
-        # While iterating through the fragments, the "carry_over_mass" keeps a track of the rejected mass difference.
-        # This is added to "diff" to get the next mass_difference.
-        carry_over_mass = 0
-
-        # "last_mass_valid" keeps a track of the last mass which was NOT rejected.
-        # This is useful for calculating the difference threshold
-        last_mass_valid = 0.0
-
-        valid_terminal_fragments = []
-        invalid_fragments = []
-        for fragment_index, (diff, mass) in enumerate(
-            zip(self.mass_diffs[side], fragment_masses)
-        ):
-            diff += carry_over_mass
-            assert pos
-
-            is_valid = True
-            if diff in self.explanations:
-                explanations = self.explanations.get(diff, [])
-            else:
-                threshold = self._calculate_diff_errors(
-                    last_mass_valid + self.mass_tags[side],
-                    last_mass_valid + diff + self.mass_tags[side],
-                    self.matching_threshold,
-                )
-                explanations = self._calculate_diff_dp(
-                    diff, threshold, modification_rate
-                )
-
-            # METHOD: if there is only one single nucleoside explanation, we can
-            # directly assign the nucleoside if there are tuples, we have to assign a
-            # possibility to the current and next position
-            # and take care of expressing that both permutations are possible
-            # Further, we consider multiple possible start positions, depending on
-            # whether the previous explanations had different lengths.
-            if explanations:
-                next_pos = set()
-                if side is Side.START:
-                    min_p = min(pos)
-                    max_p = max(pos)
-                else:
-                    min_p = max(pos)
-                    max_p = min(pos)
-                min_fragment_end = None
-                max_fragment_end = None
-
-                for p in pos:
-                    p_specific_explanations = [
-                        expl for expl in explanations if is_valid_pos(p, len(expl))
-                    ]
-                    alphabet_per_expl_len = {
-                        expl_len: set(chain(*expls))
-                        for expl_len, expls in groupby(p_specific_explanations, len)
-                    }
-
-                    if p_specific_explanations:
-                        if p == min_p:
-                            min_fragment_end = min_p + factor * min(
-                                expl_len for expl_len in alphabet_per_expl_len
-                            )
-                        elif p == max_p:
-                            max_fragment_end = max_p + factor * max(
-                                expl_len for expl_len in alphabet_per_expl_len
-                            )
-
-                    # constrain already existing sets in the range of the expl
-                    # by the nucs that are given in the explanations
-                    for expl_len, alphabet in alphabet_per_expl_len.items():
-                        for i in range(expl_len):
-                            possible_nucleosides = get_possible_nucleosides(p, i)
-
-                            if possible_nucleosides.issuperset(alphabet):
-                                # If the current explanation sharpens the list of possibilities, clear all
-                                # prior possibilities before the new explanation will add the sharpened ones below
-                                possible_nucleosides.clear()
-
-                            for j in alphabet:
-                                possible_nucleosides.add(j)
-                                # TODO: We need to do this better.
-                                # Instead of adding just the letters, we somehow need to keep a track of the possibilities to be able to constrain the LP!
-                                # We also then probably need part of the code immediately below!
-
-                    # add possible follow up positions
-                    next_pos.update(
-                        p + factor * expl_len for expl_len in alphabet_per_expl_len
-                    )
-                if max_fragment_end is None:
-                    max_fragment_end = min_fragment_end
-                if min_fragment_end is None:
-                    min_fragment_end = max_fragment_end
-                if min_fragment_end is None:
-                    # still None => both are None
-                    # TODO can we stop early in building the ladder?
-                    is_valid = False
-            elif (
-                abs(diff) <= self.matching_threshold * abs(mass + self.mass_tags[side])
-                # Problem! The above approach might blow up if the masses are very close, i.e. diff is very close to zero!
-            ):
-                if side == Side.START:
-                    min_fragment_end = min(pos)
-                    max_fragment_end = max(pos)
-                else:
-                    min_fragment_end = max(pos)
-                    max_fragment_end = min(pos)
-                next_pos = pos
-            else:
-                is_valid = False
-
-            if is_valid:
-                valid_terminal_fragments.append(
-                    TerminalFragment(
-                        index=candidate_fragments[fragment_index],
-                        min_end=min_fragment_end,
-                        max_end=max_fragment_end,
-                    )
-                )
-                pos = next_pos
-                last_mass_valid = mass
-                carry_over_mass = 0.0
-            else:
-                logger.warning(
-                    f"Skipping {side} fragment {fragment_index} with observed mass {mass} because no "
-                    "explanations are found for the mass difference."
-                )
-                carry_over_mass = diff
-
-                # Consider the skipped fragments as internal fragments! Add back the terminal mass to this fragments!
-                if fragment_index and candidate_fragments:
-                    invalid_fragments.append(candidate_fragments[fragment_index])
-
-        return skeleton_seq, valid_terminal_fragments, invalid_fragments
-
-
-class Explanation:
-    def __init__(self, *nucleosides):
-        self.nucleosides = tuple(sorted(nucleosides))
-
-    def __iter__(self):
-        yield from self.nucleosides
-
-    def __len__(self):
-        return len(self.nucleosides)
-
-    def __repr__(self):
-        return f"{{{','.join(self.nucleosides)}}}"
+# class Explanation:
+#     def __init__(self, *nucleosides):
+#         self.nucleosides = tuple(sorted(nucleosides))
+#
+#     def __iter__(self):
+#         yield from self.nucleosides
+#
+#     def __len__(self):
+#         return len(self.nucleosides)
+#
+#     def __repr__(self):
+#         return f"{{{','.join(self.nucleosides)}}}"
